requirements.txt
Django==3.2.9
python-dotenv
pandas